Get orders by user ID

+++ src/main/java/com/java/ecommerce/repositories/UserRepository.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.java.ecommerce.models.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     UserDetails findByLogin(String login);
-
 }
